requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.8.0
python-dotenv>=1.0.0 